move gol files to it's own folder
