+++ lib/game_of_life/cell.ex
@@ -12,7 +12,9 @@ defmodule GameOfLife.Cell do
     )
   end
 
-  def tick(universe_name, position), do: GenServer.call(via_tuple(universe_name, position), :tick)
+  def tick(universe_name, position) do
+    GenServer.call(via_tuple(universe_name, position), :tick)
+  end
 
   def crash(universe_name, position), do: GenServer.cast(via_tuple(universe_name, position), :crash)
 
@@ -53,15 +55,8 @@ defmodule GameOfLife.Cell do
 
   ## Utils
 
-  defp update_cell(state) do
-    alive_count = state |> neighbor_states() |> Enum.count(& &1)
-    alive = Enum.member?([2, 3], alive_count)
-
-    Map.put(state, :alive, alive)
-  end
-
-  defp neighbor_states(%{universe_name: universe_name, position: {x, y}}) do
-    [
+  defp update_cell(%{universe_name: universe_name, position: {x, y}} = state) do
+    neighbors = [
       alive?(universe_name, {x - 1, y}),
       alive?(universe_name, {x - 1, y - 1}),
       alive?(universe_name, {x, y - 1}),
@@ -71,6 +66,17 @@ defmodule GameOfLife.Cell do
       alive?(universe_name, {x, y + 1}),
       alive?(universe_name, {x + 1, y})
     ]
+
+    live_neighbor_count = Enum.count(neighbors, & &1)
+
+    alive =
+      cond do
+        live_neighbor_count < 2 -> false
+        live_neighbor_count > 3 -> false
+        true -> true
+      end
+
+    Map.put(state, :alive, alive)
   end
 
   defp exists?(name, position) do
